Move DistAbstraction and subclasses into a 'distributions' subpackage
